Resolved feedback 1.0
