+++ authentication/AuthenticationExample/Utility/Extensions.swift
@@ -16,52 +16,6 @@ import UIKit
 import Firebase
 
 
-// MARK: - Extending a `AuthProvider` to conform to `DataSourceProvidable`
-
-extension AuthProvider: DataSourceProvidable {
-    
-    private static var providers: [AuthProvider] {
-        [.Google, .Apple, .Twitter, .Microsoft, .GitHub, .Yahoo, .Facebook]
-    }
-    
-    static var providerSection: Section {
-        let providers = self.providers.map { Item(title: $0.rawValue) }
-        let header = "Identity Providers"
-        let footer = "Choose a login flow from one of the identity providers above."
-        return Section(headerDescription: header, footerDescription: footer, items: providers)
-    }
-    
-    static var emailPasswordSection: Section {
-        let image = UIImage(named: "firebaseIcon")
-        let item = Item(title: self.EmailPassword.rawValue, hasNestedContent: true, image: image)
-        let footer = "A example login flow with password authentication."
-        return Section(footerDescription: footer, items: [item])
-    }
-    
-    static var otherSection: Section {
-        let lockSymbol = UIImage.systemImage("lock.slash.fill", tintColor: .systemOrange)
-        let phoneSymbol = UIImage.systemImage("phone.fill", tintColor: .systemOrange)
-        let anonSymbol = UIImage.systemImage("questionmark.circle.fill", tintColor: .systemOrange)
-        let shieldSymbol = UIImage.systemImage("lock.shield.fill", tintColor: .systemOrange)
-        
-        let otherOptions = [
-            Item(title: self.Passwordless.rawValue, image: lockSymbol),
-            Item(title: self.PhoneNumber.rawValue, image: phoneSymbol),
-            Item(title: self.Anonymous.rawValue, image: anonSymbol),
-            Item(title: self.Custom.rawValue, image: shieldSymbol)
-        ]
-        return Section(footerDescription: "Other authentication methods.", items: otherOptions)
-    }
-    
-    static var sections: [Section] {
-        [providerSection, emailPasswordSection, otherSection]
-    }
-    
-    var sections: [Section] { AuthProvider.sections }
-
-}
-
-
 // MARK: - Extending a `Firebase User` to conform to `DataSourceProvidable`
 
 extension User: DataSourceProvidable {
@@ -107,13 +61,6 @@ extension User: DataSourceProvidable {
 }
 
 // MARK: - UIKit Extensions
-
-extension UITableView {
-    open override var showsVerticalScrollIndicator: Bool {
-        get{ false }
-        set{}
-    }
-}
 
 extension UIViewController {
