+++ authentication/AuthenticationExample/Utility/Protocols.swift
@@ -15,6 +15,7 @@
 import Foundation
 
 
+/// Delegate for signaling that a successful login with Firebase Auth has occurred
 protocol LoginDelegate: NSObject {
     func loginDidOccur()
 }
